backend/pyproject.toml
[project]
name = "libbord-backend"
version = "0.1.0"
description = "Libbord — analytics dashboard for libraries"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100
